Fix the response to the -help flag to include the Usage() message.
Before this, you would only get the Usage message if you passed in the wrong
number of arguments.  Specifically, if you passed in -help, you'd get the list
of *options*, but it wouldn't tell you the non-optional parameters needed.

Fixed this by moving the definition of the flags.Usage() var up above the call
to flag.Parse().  That way (apparently), the flag package defaults to invoking
Usage().