Add support for scripts to reply privately to the sender of a hutbot command
Looks for scripts in ./private/, runs anything matching there as well as (or
instead of) anything found in the top-devel dir.

As part of making this work, I moved essentially all of the logic about
.missing and .all out of paths/basepaths, and into CommandScript.Process().

I'm mostly confident that the end result is:

 a) Correct (aka functions the same as before)

 b) A bit simpler to understand

But a bit of review would certainly not hurt.